changement switch dump vers json

+++ app/classes/json_loader/jsonLoaderSvc.py
@@ -21,7 +21,7 @@ from app.tools.dbTools import refreshMV
 from django.conf import settings
 import json
 import os
-
+import shutil
 
 class JsonLoader(JsonLoaderABC):
     def __init__(self):
@@ -34,6 +34,7 @@ class JsonLoader(JsonLoaderABC):
         self.json_dir = getSettingValue("JSON_AUTOLOAD")
         self.archive_dir = getSettingValue("ARCHIVE_DIR")
         self.failed_dir = getSettingValue("FAILED_DIR")
+        self.waiting_dir = getSettingValue("JSON_WAITING_JSON")
 
     # ----------------
     # public methods
@@ -84,13 +85,28 @@ class JsonLoader(JsonLoaderABC):
     def succeedWorkItem(self, work_item):
         # Don't move JSON files if we are in a "dump then json mode"
         cur_poste = PosteMeteor(work_item['meteor'])
+        if not os.path.exists(self.archive_dir + "/" + work_item['meteor'] + "/"):
+            os.makedirs(self.archive_dir + "/" + work_item['meteor'] + "/")
+
+        # Reactivate wainting json files
+        if work_item['SWITCH_TO_JSON'] is not None and  work_item['SWITCH_TO_JSON'] is True:
+            files = os.listdir(self.waiting_dir + '/' + work_item['meteor'])
+
+            # Iterate over the files and copy the JSON files to the /destination directory
+            for file in files:
+                if file.endswith('.json'):
+                    source = os.path.join(self.waiting_dir + '/' + work_item['meteor'], file)
+                    destination = os.path.join(self.json_dir, file)
+                    shutil.copy(source, destination)
+            return
+
+        # Move the json file to the waiting directory
         if (cur_poste.data.load_type & Load_Type.LOAD_FROM_DUMP_THEN_JSON) == Load_Type.LOAD_FROM_DUMP_THEN_JSON:
+            os.rename(self.json_dir + "/" + work_item['f'], self.waiting_dir + "/" + work_item['meteor'] + "/" + work_item['f'])
             return
 
         # delete the file
         if hasattr(settings, 'NO_DELETE_JSON') is True and settings.NO_DELETE_JSON is True:
-            if not os.path.exists(self.archive_dir + "/" + work_item['meteor'] + "/"):
-                os.makedirs(self.archive_dir + "/" + work_item['meteor'] + "/")
             os.rename(self.json_dir + "/" + work_item['f'], self.archive_dir + "/" + work_item['meteor'] + "/" + work_item['f'])
         else:
             os.remove(self.json_dir + "/" + work_item['f'])
